Project structure cleanup. Added another module.
