+++ rest-WS/restaurant/pom.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<parent>
+		<artifactId>track-food-truck</artifactId>
+		<groupId>engineering.sonpm.trackfoodtruck</groupId>
+		<version>0.0.1-SNAPSHOT</version>
+	</parent>
+	<modelVersion>4.0.0</modelVersion>
+
+	<groupId>engineering.sonpm.trackfoodtruck</groupId>
+	<artifactId>restaurant</artifactId>
+	<version>0.0.1-SNAPSHOT</version>
+
+
+</project>
