--- rest-WS/src/main/java/engineering/sonpm/trackfoodtruck/restWS/Constants.java
@@ -1,10 +1,0 @@
-package engineering.sonpm.trackfoodtruck.restWS;
-
-/**
- * Created by Paweł Szopa on 23/10/2018
- */
-public class Constants
-{
-	public static final String USER_PATH = "/tft/user";
-	public static final String FOOD_TRUCK_PATH = "/tft/ft";
-}
