--- rest-WS/src/main/java/engineering/sonpm/trackfoodtruck/restWS/dao/FoodTruck.java
@@ -1,8 +1,0 @@
-package engineering.sonpm.trackfoodtruck.restWS.dao;
-
-/**
- * Created by Paweł Szopa on 31/10/2018
- */
-public class FoodTruck
-{
-}
