--- rest-WS/src/main/java/engineering/sonpm/trackfoodtruck/restWS/dao/User.java
@@ -1,8 +1,0 @@
-package engineering.sonpm.trackfoodtruck.restWS.dao;
-
-/**
- * Created by Paweł Szopa on 31/10/2018
- */
-public class User
-{
-}
