--- rest-WS/src/main/java/engineering/sonpm/trackfoodtruck/restWS/dto/FoodTruck.java
@@ -1,28 +1,0 @@
-package engineering.sonpm.trackfoodtruck.restWS.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-
-/**
- * Created by Paweł Szopa on 31/10/2018
- */
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class FoodTruck
-{
-	@NonNull
-	private String id;
-
-	private String name;
-
-	private String owner;
-
-	private String typeOfFood;
-
-	private String description;
-
-	private Location location;
-}
