--- rest-WS/src/test/java/engineering/sonpm/trackfoodtruck/restWS/RestWsApplicationTests.java
@@ -1,18 +1,0 @@
-package engineering.sonpm.trackfoodtruck.restWS;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-import engineering.sonpm.trackfoodtruck.restWS.application.RestWsApplication;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest(classes = RestWsApplication.class)
-public class RestWsApplicationTests
-{
-	@Test
-	public void contextLoads()
-	{
-	}
-}
