+++ rest-WS/user/src/main/java/backend/Constants.java
@@ -1,0 +1,10 @@
+package backend;
+
+/**
+ * Created by Paweł Szopa on 23/10/2018
+ */
+public class Constants
+{
+	public static final String USER_PATH = "/tft/user";
+	public static final String FOOD_TRUCK_PATH = "/tft/ft";
+}
