+++ rest-WS/user/src/main/java/backend/controller/FoodTruckService.java
@@ -1,0 +1,11 @@
+package backend.controller;
+
+import backend.dto.FoodTruck;
+
+/**
+ * Created by Paweł Szopa on 31/10/2018
+ */
+public interface FoodTruckService
+{
+	FoodTruck getFoodTruck(final String id);
+}
