+++ rest-WS/user/src/main/java/backend/controller/UserService.java
@@ -1,0 +1,13 @@
+package backend.controller;
+
+import backend.dto.User;
+
+/**
+ * Created by Paweł Szopa on 31/10/2018
+ */
+public interface UserService
+{
+	User get(final String id);
+
+	User create(final String id);
+}
