+++ rest-WS/user/src/main/java/backend/dao/User.java
@@ -1,0 +1,8 @@
+package backend.dao;
+
+/**
+ * Created by Paweł Szopa on 31/10/2018
+ */
+public class User
+{
+}
