+++ rest-WS/user/src/main/java/backend/dto/FoodTruck.java
@@ -1,0 +1,28 @@
+package backend.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+
+/**
+ * Created by Paweł Szopa on 31/10/2018
+ */
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class FoodTruck
+{
+	@NonNull
+	private String id;
+
+	private String name;
+
+	private String owner;
+
+	private String typeOfFood;
+
+	private String description;
+
+	private Location location;
+}
