+++ rest-WS/user/src/main/resources/application.properties
