+++ rest-WS/user/src/test/java/backend/RestWsApplicationTests.java
@@ -1,0 +1,18 @@
+package backend;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import backend.application.TrackFoodTruckApplication;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes = TrackFoodTruckApplication.class)
+public class RestWsApplicationTests
+{
+	@Test
+	public void contextLoads()
+	{
+	}
+}
